test(server): split public listener contracts
